split DocumentTest class
spit into a general one, one that is issue related and
one which is PDF generator related
