+++ tests/PHPUnit/Integration/DocumentTest.php
@@ -40,9 +40,11 @@ use Smalot\PdfParser\Document;
 use Smalot\PdfParser\Header;
 use Smalot\PdfParser\Page;
 use Smalot\PdfParser\Pages;
-use Smalot\PdfParser\Parser;
 use Smalot\PdfParser\PDFObject;
 
+/**
+ * General Document related tests.
+ */
 class DocumentTest extends TestCase
 {
     protected function getDocumentInstance(): Document
@@ -229,178 +231,5 @@ class DocumentTest extends TestCase
         // Missing catalog
         $document = $this->getDocumentInstance();
         $document->getPages();
-    }
-
-    /**
-     * Tests getText method without a given page limit.
-     *
-     * @see https://github.com/smalot/pdfparser/pull/562
-     */
-    public function testGetTextNoPageLimit(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/bugs/Issue331.pdf');
-
-        self::assertStringContainsString('Medeni Usul ve İcra İflas Hukuku', $document->getText());
-    }
-
-    /**
-     * Tests getText method with a given page limit.
-     *
-     * @see https://github.com/smalot/pdfparser/pull/562
-     */
-    public function testGetTextWithPageLimit(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/bugs/Issue331.pdf');
-
-        // given text is on page 2, it has to be ignored because of that
-        self::assertStringNotContainsString('Medeni Usul ve İcra İflas Hukuku', $document->getText(1));
-    }
-
-    /**
-     * Tests extraction of XMP Metadata vs. getHeader() data.
-     *
-     * @see https://github.com/smalot/pdfparser/pull/606
-     */
-    public function testExtractXMPMetadata(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/XMP_Metadata.pdf');
-
-        $details = $document->getDetails();
-
-        // Test that the dc:title data was extracted from the XMP
-        // Metadata.
-        self::assertStringContainsString("Enhance PdfParser\u{2019}s Metadata Capabilities", $details['dc:title']);
-    }
-
-    /**
-     * Tests PDFDocEncoding decode of Document Properties
-     *
-     * @see https://github.com/smalot/pdfparser/issues/609
-     */
-    public function testPDFDocEncodingDecode(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/bugs/Issue609.pdf');
-
-        $details = $document->getDetails();
-
-        // These test that Adobe-inserted \r are removed from a UTF-8
-        // escaped metadata string, and the surrounding characters are
-        // repaired
-        $testKeywords = '˘ˇˆ˙˝˛˞˜•†‡…—–ƒ⁄‹›−‰„“”‘’‚™ﬁﬂŁŒŠŸŽıłœšž€¡¢£¤¥¦§¨©ª«¬®¯°±²³´µ¶·¸¹º»¼½¾¿ÀÁÂÃÄÅÆÇÈÉÊËÌÍÎÏÐÑÒÓÔÕÖ×ØÙÚÛÜÝÞßàáâãäåæçèéêëìíîïðñòóôõö÷øùúûüýþÿ';
-        self::assertStringContainsString($testKeywords, $details['Keywords']);
-
-        $testKeywords = 'added line-feeds often destroy multibyte characters';
-        self::assertStringContainsString($testKeywords, $details['Keywords']);
-
-        // This tests that the PDFDocEncoding characters that differ
-        // from CP-1252 are decoded to their correct UTF-8 code points
-        // as well as removing \r line-feeds
-        $testSubject = '•†‡…—–ƒ⁄‹›−‰„“”‘’‚™ŁŒŠŸŽıłœšž';
-        self::assertStringContainsString($testSubject, $details['Subject']);
-    }
-
-    /**
-     * Test getText result.
-     *
-     * PDF generated with Chromium 116 via SaveAs-dialog.
-     */
-    public function testGetTextPull634Chromium(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/grouped-by-generator/R2RML-Spec_Generated_by_Chromium-SaveAs-PDF.pdf');
-
-        self::assertStringContainsString('R2RML: RDB to RDF Mapping Language', $document->getText());
-    }
-
-    /**
-     * Test getText result.
-     *
-     * PDF (1.4) generated with LibreOffice Writer (6.4).
-     *
-     * @see https://help.libreoffice.org/6.4/en-US/text/shared/01/ref_pdf_export.html
-     */
-    public function testGetTextPull634LibreOffice(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/grouped-by-generator/RichDocument_Generated_by_Libreoffice-6.4_PDF-v1.4.pdf');
-
-        self::assertStringContainsString(
-            'Some currency symbols: £, €, ¥'."\n".'German characters: ÄÖÜß',
-            $document->getText()
-        );
-    }
-
-    /**
-     * Test getText result.
-     *
-     * PDF (v 1.4) generated with Inkscape 0.92.
-     */
-    public function testGetTextPull634InkscapePDF14(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/grouped-by-generator/SimpleImage_Generated_by_Inkscape-0.92_PDF-v1.4.pdf');
-
-        self::assertEquals('TEST', $document->getText());
-    }
-
-    /**
-     * Test getText result.
-     *
-     * PDF (v 1.5) generated with Inkscape 0.92.
-     */
-    public function testGetTextPull634InkscapePDF15(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/grouped-by-generator/SimpleImage_Generated_by_Inkscape-0.92_PDF-v1.5.pdf');
-
-        self::assertEquals('TEST', $document->getText());
-    }
-
-    /**
-     * Test getText result.
-     *
-     * PDF (v 1.7) generated with Microsoft Print-to-PDF via Firefox.
-     */
-    public function testGetTextPull634MicrosoftPDF17(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/grouped-by-generator/Wikipedia-PDF_Generated_by_Microsoft_Print-to-PDF.pdf');
-
-        $outputText = $document->getText();
-
-        self::assertStringContainsString(
-            'Adobe PDF icon'."\n".'Filename'."\n".'extension',
-            $outputText
-        );
-
-        self::assertStringContainsString(
-            'are necessary to make, use, sell, and distribute PDF-compliant',
-            $outputText
-        );
-    }
-
-    /**
-     * Test getText result.
-     *
-     * PDF generated from .docx with SmallPDF (https://smallpdf.com)
-     */
-    public function testGetTextPull634SmallPDF(): void
-    {
-        $document = (new Parser())->parseFile($this->rootDir.'/samples/grouped-by-generator/Document_Generated_by_SmallPDF.pdf');
-
-        $outputText = $document->getText();
-
-        // Actual encoded spaces in the document are preserved
-        self::assertStringContainsString(
-            'SmallPDF                       SMALLPDF                             SmallPDF',
-            $outputText
-        );
-
-        // Hebrew text
-        self::assertStringContainsString(
-            'Hebrew Keyboard - תדלקמ תירבעב - Type Hebrew Online',
-            $outputText
-        );
-
-        // Russian text
-        self::assertStringContainsString(
-            'Russian Keyboard - русская клавиатура - Type Russian',
-            $outputText
-        );
     }
 }
